♻️ Move References functions into util file
