+++ src/controllers/dictionaryController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import * as dictionaryService from '../services/dictionaryService';
 import { BadRequestError } from '../utils/errors';
+import { replaceReferences } from '../utils/references';
 import { diff as diffUtil } from '../diff/DictionaryDiff';
 import logger from '../config/logger';
 
@@ -11,9 +12,7 @@ export const listDictionaries = async (req: Request, res: Response) => {
 
   if (name && version) {
     const dict = await dictionaryService.findOne(name, version);
-    const response = showReferences
-      ? dict.toObject()
-      : dictionaryService.replaceReferences(dict.toObject());
+    const response = showReferences ? dict.toObject() : replaceReferences(dict.toObject());
     res.status(200).send([response]);
   } else {
     const dicts = await dictionaryService.listAll();
@@ -27,9 +26,7 @@ export const getDictionary = async (req: Request, res: Response) => {
   const id = req.params.dictId;
 
   const dict = await dictionaryService.getOne(id);
-  const response = showReferences
-    ? dict.toObject()
-    : dictionaryService.replaceReferences(dict.toObject());
+  const response = showReferences ? dict.toObject() : replaceReferences(dict.toObject());
   res.status(200).send(response);
 };
 
@@ -59,12 +56,8 @@ export const diffDictionaries = async (req: Request, res: Response) => {
   if (name && leftVersion && rightVersion) {
     const dict1Doc = await dictionaryService.findOne(name, leftVersion);
     const dict2Doc = await dictionaryService.findOne(name, rightVersion);
-    const dict1 = showReferences
-      ? dict1Doc.toObject()
-      : dictionaryService.replaceReferences(dict1Doc.toObject());
-    const dict2 = showReferences
-      ? dict2Doc.toObject()
-      : dictionaryService.replaceReferences(dict2Doc.toObject());
+    const dict1 = showReferences ? dict1Doc.toObject() : replaceReferences(dict1Doc.toObject());
+    const dict2 = showReferences ? dict2Doc.toObject() : replaceReferences(dict2Doc.toObject());
     const diff = diffUtil(dict1, dict2);
     res.status(200).send(Array.from(diff.entries()));
   } else {
